+++ pkg/bootstrap/context.go
@@ -1,4 +1,4 @@
-package config
+package bootstrap
 
 import (
 	"context"
